first cleaning round
