+++ src/main/python/evaluation.py
@@ -45,7 +45,7 @@ Metrics:
 """
 
 
-def evaluate_intenvory(dbs, gt_path):
+def evaluate_inventory(dbs, gt_path):
     # Load JSON groundtruth
     with open(gt_path) as f:
         gt_data = json.load(f)
@@ -161,4 +161,4 @@ if __name__ == "__main__":
     # dbs += ['ALL-SIMPLE-CHIP-1', 'TEAM-PEI-1', 'TEAM-PEI-JD-1', 'TEAM-8-1']
     # dbs = ['BASELINE-1','BASELINE-2','BASELINE-3', 'BASELINE-4', 'BASELINE-5', 'BASELINE-6', 'BASELINE-7','BASELINE-8','BASELINE-10','BASELINE-11', 'BASELINE-12', 'BASELINE-13', 'BASELINE-14', 'BASELINE-16', 'BASELINE-20', 'BASELINE-22', 'BASELINE-23', 'BASELINE-25']
     gt_path = "src/main/resources/ground_truth/v14.json"  # list of ground truth W.R.T the previous databases
-    evaluate_intenvory(dbs, gt_path)
+    evaluate_inventory(dbs, gt_path)
